Collapse Usecase hierarchy: Usecase (base) + BackendUsecase (public)
Drops the BaseUsecase + Usecase + InternalUsecase three-class split in favor
of a two-class hierarchy: Usecase is the shared base (helpers, abstract exec);
BackendUsecase extends it with the abstract argumentsSchema/resultSchema for
RPC-boundary validation. Internal-only usecases (AssistantsProcessConversation,
CollectionsDownSync) extend Usecase directly, no placeholder schemas needed.
