+++ packages/core/executing-backend/src/usecases/apps/Delete.ts
@@ -12,7 +12,7 @@ import {
 } from "@superego/shared-utils";
 import * as v from "valibot";
 import makeResultError from "../../makers/makeResultError.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import {
   appNotFound,
   commandConfirmationNotValid,
@@ -21,7 +21,9 @@ import {
 import { appId } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class AppsDelete extends Usecase<Backend["apps"]["delete"]> {
+export default class AppsDelete extends BackendUsecase<
+  Backend["apps"]["delete"]
+> {
   argumentsSchema = v.tuple([appId(), v.string()]);
   resultSchema = makeResultSchema(v.null(), [
     appNotFound(),
