+++ packages/core/executing-backend/src/usecases/apps/List.ts
@@ -5,12 +5,12 @@ import * as v from "valibot";
 import type AppVersionEntity from "../../entities/AppVersionEntity.js";
 import makeApp from "../../makers/makeApp.js";
 import assertAppVersionExists from "../../utils/assertAppVersionExists.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { app } from "../../validation/domain/app.js";
 import { unexpectedError } from "../../validation/errors.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class AppsList extends Usecase<Backend["apps"]["list"]> {
+export default class AppsList extends BackendUsecase<Backend["apps"]["list"]> {
   argumentsSchema = v.tuple([]);
   resultSchema = makeResultSchema(v.array(app()), [unexpectedError()]);
