+++ packages/core/executing-backend/src/usecases/assistants/DeleteConversation.ts
@@ -12,7 +12,7 @@ import {
 } from "@superego/shared-utils";
 import * as v from "valibot";
 import makeResultError from "../../makers/makeResultError.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import {
   commandConfirmationNotValid,
   conversationNotFound,
@@ -21,7 +21,7 @@ import {
 import { conversationId } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class AssistantsDeleteConversation extends Usecase<
+export default class AssistantsDeleteConversation extends BackendUsecase<
   Backend["assistants"]["deleteConversation"]
 > {
   argumentsSchema = v.tuple([conversationId(), v.string()]);
