+++ packages/core/executing-backend/src/usecases/assistants/GetLiveConversation.ts
@@ -7,13 +7,13 @@ import type {
 import type { ResultPromise } from "@superego/global-types";
 import { makeSuccessfulResult } from "@superego/shared-utils";
 import * as v from "valibot";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { conversation as conversationSchema } from "../../validation/domain/conversation.js";
 import { unexpectedError } from "../../validation/errors.js";
 import { conversationId } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class AssistantsGetLiveConversation extends Usecase<
+export default class AssistantsGetLiveConversation extends BackendUsecase<
   Backend["assistants"]["getLiveConversation"]
 > {
   argumentsSchema = v.tuple([conversationId()]);
