+++ packages/core/executing-backend/src/usecases/assistants/StartConversation.ts
@@ -27,11 +27,11 @@ import type FileEntity from "../../entities/FileEntity.js";
 import UnexpectedAssistantError from "../../errors/UnexpectedAssistantError.js";
 import makeConversation from "../../makers/makeConversation.js";
 import makeResultError from "../../makers/makeResultError.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import ConversationUtils from "../../utils/ConversationUtils.js";
 import difference from "../../utils/difference.js";
 import isEmpty from "../../utils/isEmpty.js";
 import MessageContentFileUtils from "../../utils/MessageContentFileUtils.js";
-import Usecase from "../../utils/Usecase.js";
 import { conversation as conversationSchema } from "../../validation/domain/conversation.js";
 import {
   filesNotFound,
@@ -42,7 +42,7 @@ import looseObjectAs from "../../validation/helpers/looseObjectAs.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 import CollectionsList from "../collections/List.js";
 
-export default class AssistantsStartConversation extends Usecase<
+export default class AssistantsStartConversation extends BackendUsecase<
   Backend["assistants"]["startConversation"]
 > {
   argumentsSchema = v.tuple([
