+++ packages/core/executing-backend/src/usecases/boutique/GetPack.ts
@@ -13,13 +13,13 @@ import {
 } from "@superego/shared-utils";
 import * as v from "valibot";
 import makeResultError from "../../makers/makeResultError.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { pack } from "../../validation/domain/pack.js";
 import { packNotFound, unexpectedError } from "../../validation/errors.js";
 import { packId } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class BoutiqueGetPack extends Usecase<
+export default class BoutiqueGetPack extends BackendUsecase<
   Backend["boutique"]["getPack"]
 > {
   argumentsSchema = v.tuple([packId()]);
