+++ packages/core/executing-backend/src/usecases/collection-categories/Update.ts
@@ -20,7 +20,7 @@ import type CollectionCategoryEntity from "../../entities/CollectionCategoryEnti
 import makeCollectionCategory from "../../makers/makeCollectionCategory.js";
 import makeResultError from "../../makers/makeResultError.js";
 import makeValidationIssues from "../../makers/makeValidationIssues.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { collectionCategory } from "../../validation/domain/collectionCategory.js";
 import {
   collectionCategoryIconNotValid,
@@ -34,7 +34,7 @@ import { collectionCategoryId } from "../../validation/helpers/idSchemas.js";
 import looseObjectAs from "../../validation/helpers/looseObjectAs.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class CollectionCategoriesUpdate extends Usecase<
+export default class CollectionCategoriesUpdate extends BackendUsecase<
   Backend["collectionCategories"]["update"]
 > {
   argumentsSchema = v.tuple([
