+++ packages/core/executing-backend/src/usecases/collections/GetVersion.ts
@@ -14,7 +14,7 @@ import {
 import * as v from "valibot";
 import makeCollectionVersion from "../../makers/makeCollectionVersion.js";
 import makeResultError from "../../makers/makeResultError.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { collectionVersion as collectionVersionSchema } from "../../validation/domain/collection.js";
 import {
   collectionVersionNotFound,
@@ -26,7 +26,7 @@ import {
 } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class CollectionsGetVersion extends Usecase<
+export default class CollectionsGetVersion extends BackendUsecase<
   Backend["collections"]["getVersion"]
 > {
   argumentsSchema = v.tuple([
