+++ packages/core/executing-backend/src/usecases/collections/SetRemote.ts
@@ -28,7 +28,7 @@ import makeCollection from "../../makers/makeCollection.js";
 import makeResultError from "../../makers/makeResultError.js";
 import makeValidationIssues from "../../makers/makeValidationIssues.js";
 import assertCollectionVersionExists from "../../utils/assertCollectionVersionExists.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { collection as collectionDomainSchema } from "../../validation/domain/collection.js";
 import {
   cannotChangeCollectionRemoteConnector,
@@ -44,7 +44,7 @@ import { collectionId as collectionIdSchema } from "../../validation/helpers/idS
 import looseObjectAs from "../../validation/helpers/looseObjectAs.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class CollectionsSetRemote extends Usecase<
+export default class CollectionsSetRemote extends BackendUsecase<
   Backend["collections"]["setRemote"]
 > {
   argumentsSchema = v.tuple([
