+++ packages/core/executing-backend/src/usecases/collections/TriggerDownSync.ts
@@ -22,7 +22,7 @@ import makeCollection from "../../makers/makeCollection.js";
 import makeResultError from "../../makers/makeResultError.js";
 import assertCollectionRemoteConnectorExists from "../../utils/assertCollectionRemoteConnectorExists.js";
 import assertCollectionVersionExists from "../../utils/assertCollectionVersionExists.js";
-import Usecase from "../../utils/Usecase.js";
+import BackendUsecase from "../../utils/BackendUsecase.js";
 import { collection as collectionDomainSchema } from "../../validation/domain/collection.js";
 import {
   collectionHasNoRemote,
@@ -34,7 +34,7 @@ import {
 import { collectionId as collectionIdSchema } from "../../validation/helpers/idSchemas.js";
 import makeResultSchema from "../../validation/helpers/makeResultSchema.js";
 
-export default class CollectionsTriggerDownSync extends Usecase<
+export default class CollectionsTriggerDownSync extends BackendUsecase<
   Backend["collections"]["triggerDownSync"]
 > {
   argumentsSchema = v.tuple([collectionIdSchema()]);
